aqa_python/automation_api/requirements.txt
httpx==0.28.1
pytest==8.3.5
python-dotenv==1.1.0
loguru==0.7.3